cleaned up the code and moved routes in separate file
